pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "valocoach"
version = "0.1.0"
description = "Valorant tactical coaching CLI"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.4.0",
    "litellm>=1.50.0",
    "python-dotenv>=1.0.0",
    "tomli-w>=1.0.0",
    # HenrikDev async HTTP client.
    "httpx==0.27.2",
    # Dialect-agnostic ORM and async session support for SQLite now, PostgreSQL later.
    "sqlalchemy[asyncio]==2.0.49",
    # Async SQLite driver used by SQLAlchemy's async engine.
    "aiosqlite==0.22.1",
    # Schema migrations for the SQLite-to-PostgreSQL path.
    "alembic==1.18.4",
    # Retry/backoff policy for HenrikDev rate limits and transient server errors.
    "tenacity==9.1.4",
    # week 4 — vector RAG pipeline
    "chromadb>=0.5.0",
    "trafilatura>=1.12.0",
    "beautifulsoup4>=4.12.0",
    "ollama>=0.3.0",
    "tiktoken>=0.7.0",
    "youtube-transcript-api>=0.6.0",
    # week 5 — interactive REPL
    "prompt-toolkit>=3.0.50",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    # Mock HenrikDev HTTP requests without live API calls.
    "pytest-httpx==0.34.0",
    "ruff>=0.6.0",
    # LLM eval harness (scripts/run_eval.py)
    "pyyaml>=6.0.0",
]

[project.scripts]
valocoach = "valocoach.cli.app:app"

[tool.hatch.build.targets.wheel]
packages = ["src/valocoach"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM", "RUF"]
ignore = ["E501", "RUF001"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
